[test] apply the new test style to databend-query - sql (132/177)
