+++ query/tests/it/sql/statements/query/query_normalizer.rs
@@ -15,10 +15,10 @@
 use common_base::tokio;
 use common_exception::ErrorCode;
 use common_exception::Result;
+use databend_query::sql::statements::query::QueryNormalizer;
+use databend_query::sql::DfParser;
+use databend_query::sql::DfStatement;
 
-use crate::sql::statements::query::QueryNormalizer;
-use crate::sql::DfParser;
-use crate::sql::DfStatement;
 use crate::tests::create_query_context;
 
 #[tokio::test]
